+++ src/Models/KeystoneRole.php
@@ -2,8 +2,16 @@
 
 namespace BSPDX\Keystone\Models;
 
+use Illuminate\Support\Collection;
 use Spatie\Permission\Models\Role;
 
+/**
+ * @method $this givePermissionTo(...$permissions)
+ * @method $this syncPermissions(...$permissions)
+ * @method $this revokePermissionTo($permission)
+ * @method bool hasPermissionTo($permission, ?string $guardName = null)
+ * @method Collection getPermissionNames()
+ */
 class KeystoneRole extends Role
 {
     /**